Added field to authenticate_form convenience function to add easy key.
This form field allows the addon to put in the public_key while they are
logging in. If the site supports this, the public_key can be associated
with the account when a user logs in with his normal username and
password combo. This makes it easy for users to start using TrustAuth.